backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
python-multipart
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
bcrypt==3.2.2
